Fix error if no matching packages found
In the scenario where a project is using prerelease packages but specifies `--pre-release:Never` and there are no stable packages available, dotnet outdated would return an error and list blank versions to upgrade to. Instead, if no packages are found for upgrade, report the package itself as the latest version.