Update package removal condition before upgrade (#605)
The code now includes an additional check to determine if the package version is centrally managed (`package.IsVersionCentrallyManaged`). If the project is not SDK style or the package version is centrally managed, the package will be removed before the upgrade. This fixes an issue with upgrading legacy projects using CPM, as the reference cannot be automatically removed from the central file.